[web\schema] rename module
