+++ src/web/schemas/__init__.py
@@ -1,2 +1,2 @@
-from src.web.schemas.poem import *  # noqa: F403
+from src.web.schemas.poem_result import *  # noqa: F403
 from src.web.schemas.poem_forms import *  # noqa: F403
